check for params

+++ bin/basenji_train_hic.py
@@ -75,7 +75,8 @@ def main(_):
   import shutil
   if not os.path.isdir(FLAGS.log_dir):
     os.mkdir(FLAGS.log_dir)
-  shutil.copy(FLAGS.params,FLAGS.log_dir+'/params.json')
+  if not os.path.isfile(FLAGS.log_dir+'/params.json'):
+    shutil.copy(FLAGS.params,FLAGS.log_dir+'/params.json')
 
   import tensorflow as tf
   if tf.__version__[0] == '1':
